Add framing errors + Cleanup errors
